Refactor table sort code
